+++ py2app/bootstrap/boot_aliasplugin.py
@@ -1,6 +1,23 @@
+import re, sys
+cookie_re = re.compile(b"coding[:=]\s*([-\w.]+)")
+if sys.version_info[0] == 2:
+    default_encoding = 'ascii'
+else:
+    default_encoding = 'utf-8'
+
+def guess_encoding(fp):
+    for i in range(2):
+        ln = fp.readline()
+
+        m = cookie_re.search(ln)
+        if m is not None:
+            return m.group(1).decode('ascii')
+
+    return default_encoding
+
 def _run():
     global __file__
-    import os, sys, site
+    import os, site
     sys.frozen = 'macosx_plugin'
 
     if 'ARGVZERO' in os.environ:
@@ -10,9 +27,14 @@ def _run():
     script = SCRIPT_MAP.get(argv0, DEFAULT_SCRIPT)
 
     sys.argv[0] = __file__ = script
-    with open(script, 'rU') as fp:
-        source = fp.read() + "\n"
+    if sys.version_info[0] == 2:
+        with open(script, 'rU') as fp:
+            source = fp.read() + "\n"
+    else:
+        with open(script, 'rb') as fp:
+            encoding = guess_encoding(fp)
+
+        with open(script, 'r', encoding=encoding) as fp:
+            source = fp.read() + '\n'
 
     exec(compile(source, script, 'exec'), globals(), globals())
-
-