+++ py2app/bootstrap/boot_plugin.py
@@ -1,6 +1,23 @@
+import re, sys
+cookie_re = re.compile(b"coding[:=]\s*([-\w.]+)")
+if sys.version_info[0] == 2:
+    default_encoding = 'ascii'
+else:
+    default_encoding = 'utf-8'
+
+def guess_encoding(fp):
+    for i in range(2):
+        ln = fp.readline()
+
+        m = cookie_re.search(ln)
+        if m is not None:
+            return m.group(1).decode('ascii')
+
+    return default_encoding
+
 def _run():
     global __file__
-    import os, sys, site
+    import os, site
     sys.frozen = 'macosx_plugin'
     base = os.environ['RESOURCEPATH']
 
@@ -10,9 +27,15 @@ def _run():
         argv0 = None
     script = SCRIPT_MAP.get(argv0, DEFAULT_SCRIPT)
 
-    path = os.path.join(base, script)
-    __file__ = path
-    with open(path, 'rU') as fp:
-        source = fp.read()
-    exec(compile(source, path, 'exec'), globals(), globals())
+    __file__ = path = os.path.join(base, script)
+    if sys.version_info[0] == 2:
+        with open(script, 'rU') as fp:
+            source = fp.read() + "\n"
+    else:
+        with open(script, 'rb') as fp:
+            encoding = guess_encoding(fp)
 
+        with open(script, 'r', encoding=encoding) as fp:
+            source = fp.read() + '\n'
+
+    exec(compile(source, path, 'exec'), globals(), globals())