+++ py2app_tests/basic_app_with_encoding/package3/mod.py
@@ -1,0 +1,1 @@
+""" package3.mod """